vendor_platform/api/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
pydantic==2.6.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
bcrypt==4.0.1
